HTM-733 Remove the voice receiver from mockorbcallbacksession
Description:
- The receiver should not process voice command through orbsessioncallback
- Send data directly to orbsession

Testing:
- Voice command
